+++ validator_client/src/http_api/sign_voluntary_exit.rs
